+++ src/component/pages/RouterComponent/RouterComponent.js
@@ -16,13 +16,13 @@ import Year1958 from '../Year1958/Year1958';
 import Year1959 from '../Year1959/Year1959';
 import Year1960 from '../Year1960/Year1960';
 import Year1961 from '../Year1961/Year1961';
-import DefaultLayout from '../DefaultLayout/DefaultLayout';
+import Timeline from '../Timeline/Timeline';
 
 
 export default function RouterComponent () {
     return (
         <Router>
-            <DefaultLayout /> 
+            <Timeline /> 
             <Switch>
                 <Route exact path="/" component={Year1950} />
                 <Route exact path="/1950" component={Year1950} />
